GEODE-6096: Extract DUnit inner classes
